feat(evm): implement CALL opcode family, fix calculation of the `call_id` for interrupted calls, fix running rwasm e2e tests by adding a new flag, recalculate gas for e2e tests since gas policy has changed, simplify interruption outcome structure, fix processing outcomes for nested calls, fix gas calculation for interrupted calls, fix eip-7702 gas calculation for gas free contracts

+++ crates/types/src/address.rs
@@ -209,8 +209,8 @@ mod tests {
             _call_id: u32,
             _return_data: &[u8],
             _exit_code: i32,
-            _fuel_used: u64,
-        ) -> (u64, i32) {
+            _fuel_used: u32,
+        ) -> (u32, i32) {
             todo!()
         }
